+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 - New logger to send events to New Relic as NR-Logs.
+- New logger to send events in Tracy bar.
 - New logger to send metrics to New Relic.
 - New logger to send traces to New Relic.
 - New PerfOps One menu in the admin bar for quick actions.